Asteroids 0.0.7

+++ GameTemplate/Vector2.cpp
@@ -1,7 +1,7 @@
 #include "Vector2.h"
 #include <math.h>
 
-Vector2::Vector2(int x, int y) {
+Vector2::Vector2(float x, float y) {
 	this->x = x;
 	this->y = y;
 }
